src/server/message.ts
import type { Upstream } from "../protocol/up.js";
import type { ClientID, ClientMap, Socket } from "../types/client-state.js";
import type { LogContext } from "../util/logger.js";
import { sendError } from "../util/socket.js";
import { handlePush, type ProcessUntilDone } from "./push.js";
import { handlePing } from "./ping.js";

/**
 * Handles an upstream message coming into the server by dispatching to the
 * appropriate handler.
 */
export function handleMessage(
  lc: LogContext,
  clientMap: ClientMap,
  clientID: ClientID,
  data: string,
  ws: Socket,
  processUntilDone: ProcessUntilDone
) {
  let message;
  try {
    message = getMessage(data);
  } catch (e) {
    lc.info?.("invalid message", e);
    sendError(ws, String(e));
    return;
  }

  const client = clientMap.get(clientID);
  if (!client) {
    lc.error?.("client not found, closing socket");
    sendError(ws, `no such client: ${clientID}`);
    // This is not expected to ever occur.  However if it does no pushes will
    // ever succeed over this connection since it is missing an entry in
    // ClientMap.  Close connection so client can try to reconnect and recover.
    ws.close();
    return;
  }

  switch (message[0]) {
    case "ping":
      handlePing(lc, ws);
      break;
    case "push":
      handlePush(lc, client, message[1], () => Date.now(), processUntilDone);
      break;
    default:
      throw new Error(`Unknown message type: ${message[0]}`);
  }
}

function getMessage(data: string): Upstream {
  const json = JSON.parse(data);
  //return upstreamSchema.parse(json);
  return json as Upstream;
}
